indexjs分割完了
