+++ vue/src/views/Chart.vue
@@ -78,14 +78,31 @@ export default {
     this.renderChart(this.data, this.options)
   },
   methods: {
-    getAge () {
-      const age = this.$store.state.age
+    async getAge () {
+      // こっちはなんかうまくいかないw
+      // const age = this.$store.state.chart.age
+      const age = this.$store.state.chart.age
       this.data.labels = age
     },
     getLifeScores () {
-      const lifeScores = this.$store.state.lifeScores
+      const lifeScores = this.$store.state.chart.lifeScores
       this.data.datasets[0].data = lifeScores
     }
   }
+  // mounted () {
+  //   this.getAge()
+  //   this.getLifeScores()
+  //   this.renderChart(this.data, this.options)
+  // },
+  // methods: {
+  //   getAge () {
+  //     const age = this.$store.state.age
+  //     this.data.labels = age
+  //   },
+  //   getLifeScores () {
+  //     const lifeScores = this.$store.state.lifeScores
+  //     this.data.datasets[0].data = lifeScores
+  //   }
+  // }
 }
 </script>
